sqlite adapter dispatcher refactor
